Rename PluginSpecification to BasePluginSpecification
